old-tests/submission/Microsoft/selection/setSelectionRange.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Call setSelectionRange() on a text field</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var input1 = document.getElementById("input1");
                var expectedResult = "input";

                input1.setSelectionRange(16, 21);

                checkDefaultSelectionAttributes();

                var selectionText = input1.value.substring(input1.selectionStart, input1.selectionEnd);
                assert_equals(selectionText, expectedResult);
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <input style="WIDTH: 500px" id="input1" value="Some text in an input control" type="text" />
        <p>Call setSelectionRange() on the input element to select some of the text</p>
    </body>
</html>
